pyproject.toml
[tool.poetry]
name = "optimos_v2"
version = "0.0.1"
description = "optimos_v2"
authors = ["Jannis Rosenbaum <jannis@ut.ee>"]
readme = "README.md"
package-mode = true
packages = [{ include = "o2" }]

[tool.poetry.dependencies]
python = ">=3.10,<3.11"
numpy = "^1.26.4"
# pix-framework = { path = "../pix-framework", develop = true }
prosimos = { git = "https://github.com/sinnaj-r/Prosimos.git", branch = "main" }
# prosimos = { path = "../Prosimos", develop = true }
click = "^8.1.7"
pandas = "^2.2.3"
dataclass-wizard = "^0.22.3"
sympy = "^1.12.1"
scipy = "^1.13.0"
fastapi = "^0.114.0"
uvicorn = "^0.30.6"
rtree = "^1.3.0"
xxhash = "^3.5.0"
torch = "^2.5.0"
sb3-contrib = "^2.3.0"
tensorflow = "^2.18.0"


[tool.poetry.group.dev.dependencies]
pytest-cov = "^5.0.0"
pytest-watch = "^4.2.0"
tensorboard = "^2.18.0"
memray = "^1.14.0"

[tool.poetry.scripts]
server = "o2_server.server:start"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.coverage.run]
source = ["o2"]

[tool.pytest.ini_options]
addopts = "--ignore=states --ignore=evaluations"
# addopts = "--cov --cov-report=lcov:lcov.info --cov-report=term"
pythonpath = ["."]

[tool.ruff]
lint.select = [
    "E",   # pycodestyle
    "F",   # Pyflakes
    "UP",  # pyupgrade
    "B",   # flake8-bugbear
    "SIM", # flake8-simplify
    "I",   # isort
    "D",   # Docstring conventions (pydocstyle)
    "N",   # PEP 8 naming conventions
    "W",   # Pycodestyle warnings
    "B",   # Bugbear
    "Q",   # Flake8-quotes
    "ANN", # Annotations (flake8-annotations)
]
lint.ignore = [
    "D100",   # Missing docstring in public module
    "D107",
    "ANN202",
    "D203",
    "D213",
]
line-length = 110


[tool.ruff.lint.per-file-ignores]
"tests/*" = ["D", "ANN"]
